# docs/04-API명세-openapi.yaml
openapi: 3.0.3
info:
  title: HOOPLY API
  version: 1.0.0
  description: 농구 동호회 팀 매칭 서비스. 인증은 Bearer JWT(access 30분/refresh 14일). 모든 4xx/5xx 본문은 ErrorResponse {code, message, details[]}.
servers:
- url: https://hooply-backend.onrender.com/api/v1
  description: 운영
- url: http://localhost:8000/api/v1
  description: 로컬
security:
- bearerAuth: []
paths:
...
  /teams:
    post:
      summary: 팀 생성
      tags:
      - 팀 · 참가자
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamCreated'
  /teams/join:
    post:
      summary: 팀 코드로 가입
      tags:
      - 팀 · 참가자
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamJoinRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
        '403':
          description: REMOVED_FROM_TEAM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: TEAM_CODE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: ALREADY_MEMBER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{team_id}:
    get:
      summary: 팀 상세 조회
      tags:
      - 팀 · 참가자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
        '403':
          description: NOT_A_MEMBER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: 팀 정보 수정
      tags:
      - 팀 · 참가자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpdate'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{team_id}/code:regenerate:
    post:
      summary: 팀 코드 재발급
      tags:
      - 팀 · 참가자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamCodeView'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{team_id}/players:
    get:
      summary: 팀원 목록 조회
      tags:
      - 팀 · 참가자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      - name: kind
        in: query
        required: false
        schema:
          nullable: true
      - name: status
        in: query
        required: false
        schema:
          nullable: true
      - name: sort
        in: query
        required: false
        schema:
          type: string
          enum:
          - skill
          - name
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemList_Union_PlayerCard__PlayerCardDetailed__'
  /teams/{team_id}/players/{player_id}/role:
    patch:
      summary: 매니저 권한 부여/회수
      tags:
      - 팀 · 참가자
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerCard'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: CANNOT_DEMOTE_LAST_MANAGER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{team_id}:leave:
    post:
      summary: 팀 나가기
      tags:
      - 팀 · 참가자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: 성공
        '422':
          description: CANNOT_DEMOTE_LAST_MANAGER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{team_id}/players/{player_id}:
    delete:
      summary: 팀원 제외
      tags:
      - 팀 · 참가자
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: 성공
        '403':
          description: FORBIDDEN_ROLE / FORBIDDEN_NOT_OWNER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: CANNOT_DEMOTE_LAST_MANAGER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorDetail:
      properties:
        field:
          type: string
          nullable: true
        reason:
          type: string
        context:
          additionalProperties: true
          type: object
          nullable: true
      type: object
      required:
      - reason
    ErrorResponse:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
      type: object
      required:
      - code
      - message
...
    ItemList_Union_PlayerCard__PlayerCardDetailed__:
      properties:
        items:
          items:
            anyOf:
            - $ref: '#/components/schemas/PlayerCard'
            - $ref: '#/components/schemas/PlayerCardDetailed'
          type: array
      type: object
      required:
      - items
...
    PlayerCard:
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        kind:
          type: string
        display_name:
          type: string
        role:
          type: string
        profile_image_url:
          type: string
          nullable: true
        height_cm:
          type: integer
          nullable: true
        skill_grade:
          $ref: '#/components/schemas/SkillGrade'
          nullable: true
        primary_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
        attendance_rate:
          type: number
          nullable: true
        skill_confidence:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
      type: object
      required:
      - id
      - kind
      - display_name
      - role
    PlayerCardDetailed:
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        kind:
          type: string
        display_name:
          type: string
        role:
          type: string
        profile_image_url:
          type: string
          nullable: true
        height_cm:
          type: integer
          nullable: true
        skill_grade:
          $ref: '#/components/schemas/SkillGrade'
          nullable: true
        primary_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
        attendance_rate:
          type: number
          nullable: true
        skill_confidence:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        skill_overall:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        prior_overall:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        skill_axes:
          additionalProperties:
            type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
            nullable: true
          type: object
        avg_margin:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        quarters_played:
          type: integer
        attended_events:
          type: integer
      type: object
      required:
      - id
      - kind
      - display_name
      - role
...
    Position:
      type: string
      enum:
      - PG
      - SG
      - SF
      - PF
      - C
...
    RoleUpdate:
      properties:
        role:
          $ref: '#/components/schemas/TeamRole'
      type: object
      required:
      - role
...
    SkillGrade:
      type: string
      enum:
      - A
      - B
      - C
      - D
      - E
...
    TeamCodeView:
      properties:
        team_code:
          type: string
      type: object
      required:
      - team_code
    TeamCreate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
        home_court:
          type: string
          maxLength: 100
          nullable: true
      type: object
      required:
      - name
    TeamCreated:
      properties:
        id:
          type: integer
        team_code:
          type: string
      type: object
      required:
      - id
      - team_code
    TeamDetail:
      properties:
        id:
          type: integer
        name:
          type: string
        team_code:
          type: string
        status:
          type: string
        approval_status:
          type: string
        min_members:
          type: integer
        home_court:
          type: string
          nullable: true
        owner:
          $ref: '#/components/schemas/UserSummary'
        member_count:
          type: integer
        my_role:
          $ref: '#/components/schemas/TeamRole'
          nullable: true
        my_player_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
      type: object
      required:
      - id
      - name
      - description
      - team_code
      - status
      - min_members
      - home_court
      - owner
      - member_count
      - created_at
    TeamJoinRequest:
      properties:
        team_code:
          type: string
          maxLength: 8
          minLength: 8
      type: object
      required:
      - team_code
...
    TeamRole:
      type: string
      enum:
      - MANAGER
      - PLAYER
    TeamUpdate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          nullable: true
        home_court:
          type: string
          maxLength: 100
          nullable: true
      type: object
...
    UserSummary:
      properties:
        id:
          type: integer
        name:
          type: string
        nickname:
          type: string
          nullable: true
        profile_image_url:
          type: string
          nullable: true
      type: object
      required:
      - id
      - name
...